refactor: extract shared sim geometry modules
Split the convention-carrying geometry out of the sim renderers into
neutral modules both simulator sides may import: ellipsoid normals and
Lambert shading (ellipsoid_geometry), the polyhedral-mesh primitives
(mesh_geometry, renamed from sim_body_polyhedral), and the mode-1 ring
edge math with the border-atop rasterizer (ring_geometry).  Shared
helpers take explicit geometry arguments and never read the scene
mapping, so the image-side renderer and the navigator-side predicted
models place features with identical conventions without exchanging
scene information.
